Fix ambiguities in reading schema json
Dataset.read() is now stricter about the json schema, raising an error if an
acceptable type is not found. (However, for the case where the representation is
given as a list of string, only a check that the representation is a list is
made, the type of the elements is not checked. Furthermore, no attempt is made
to validate the values read against the types that they are claimed to
be. (Except that an attempt to read string as a numeric type, or something that
cannot be parsed as a date or datetime as a date or datetime, will result in a
runtime error.)

Fixes issues #25 and #26