Dataset schema (#33)
* Change version requirements for matplotlib

From =3.4.3 to >=3.4.3

* Start definition of table schema

* Rewrite documents for schema

* Initial complete version of dataset schema format

* Add new json table description for Texas data

* Move schema documentation from markdown to ReST

* census teaching file schema (#23)

merging since it doesn't break anything

* adding type mapping

* adding type mapping into read csv

* removing row column

* adding CategoricalDtype example

* Fix table in schema documentation

* Add dataset schema documentation to docs/source/

* Update index.rst

* Minor edits to dataset documentation

* Add `*~` (emacs autosave files) to gitignore

* Minor edits to TabularDataset

* Fix ambiguities in reading schema json

Dataset.read() is now stricter about the json schema, raising an error if an
acceptable type is not found. (However, for the case where the representation is
given as a list of string, only a check that the representation is a list is
made, the type of the elements is not checked. Furthermore, no attempt is made
to validate the values read against the types that they are claimed to
be. (Except that an attempt to read string as a numeric type, or something that
cannot be parsed as a date or datetime as a date or datetime, will result in a
runtime error.)

Fixes issues #25 and #26

* Fix error in spelling of "it's"

Co-authored-by: Callum Mole <cmole@turing.ac.uk>
Co-authored-by: crangelsmith <crangelsmith@turing.ac.uk>
Co-authored-by: James Jordon <jjordon95@gmail.com>